requirements.txt
asgiref==3.3.1
Django==3.1.7
paho-mqtt==1.5.1
pytz==2021.1
sqlparse==0.4.1
